Refactor duplicated S3 testcontainer access interfaces
